screenshot: extract screenshot.go
Reasons:

- CaptureScreenshot has nothing to do with navigation, so it's ok
to move it out of the nav.go file.
- FullScreenshot does not use emulate action any more, so it's ok
to move it out of the emulate.go file.
- query.go is a little big and I'm planning to add more unit test
for the screenshot feature, so it's better to move the feature to
another file.

This change is a simple refactor, and does not change any behavior
of the source code.
